Azure backend support (#322)
* Implement Azure backend support

This is a summary of our modifications to the official project:
* Added an Azure version of sample xgboost-high-level-apis-azure.ipynb
* Added a KubeflowAzureBackend
* Added an Azure Storage based context
* Added Azure related helper functions and classes
* Added Azure related constants
* Added Azure library to requirements

In fairing/constants/constants.py you can find the secrets required for
the code to work and how to create them.

* Unit test for Azure credentials

* Fix service_type error

The error got introduced with the commit that allows users to
set the service_type via the API.

As a fix, getting rid of the overwritten functions in AzureBackend
because they are equal to the ones in the KubernetesBackend (parent).

* Fix local variable reference

The local variable `tail` was referenced before assignment in
certain code paths.

* Integration test for prediction with Azure backend

* Allow specifying which tests to run

Changing the syntax in the entrypoint keeps the original default
behaviour but also allows specifying which tests to run. This could
be used to run only a specific integration test or exclude some
folders from the test run.

* Code cleanup and an additional test

Cleanup mostly based on pylint recommendations.

* Align environment variable names

* Ignore Azure integration tests in Cloud Build

* Align secret names with Kubeflow scripts

The name `azcreds` was used as the name of the secret at
https://github.com/kubeflow/kubeflow/blob/2f94612ce62eaa2a5ba23da7c0515b65dac380a7/scripts/azure/util.sh#L100

This change aligns with the script above in that if secrets are set
that way, they will be picked up correctly.

* Refactor to a single sample using many backends

To avoid code duplication, the Azure-specific sample notebook was
removed and the existing sample was modified to contain changes
required to use different backends.

* Skipping GCPManagedBackend test

* Impoting pytest

* updating the correct test this time and enabling only one test file

* enabling all tests

* Fix pylint issues